fix JPS err, apply it to Enemy

+++ Assets/Scripts/EnemyUI/JPS/JPSNode.cs
@@ -1,5 +1,6 @@
 using System.Collections;
 using System.Collections.Generic;
+using System.Data;
 using UnityEngine;
 
 namespace JPS {
@@ -17,11 +18,11 @@ namespace JPS {
         public JPSNode parent;
         public Vector2Int pos;
         public JPSDir dir;
-        private float cost;
-        private float heuri;
+        private int cost;
+        private int heuri;
 
 
-        public JPSNode(JPSNode parent, Vector2Int pos, JPSDir dir, float cost, float huri)
+        public JPSNode(JPSNode parent, Vector2Int pos, JPSDir dir, int cost, int huri)
         {
             this.parent = parent;
             this.pos = pos;
@@ -30,7 +31,7 @@ namespace JPS {
             this.heuri = huri;
         }
 
-        public float GetExpectedCost()
+        public int GetExpectedCost()
         {
             return cost + heuri;
         }
